Add files via upload
This is my first time using sci-kit learn to fit a model to a very simple dataset. We plot the data and upon doing this we clearly see a linear relationship between the two variables. Using the linear regression model in sklearn we find the precise line of best for the data.